// frontend/src/api/types/com/whtwnd/blog/getAuthorPosts.ts
/**
 * GENERATED CODE - DO NOT MODIFY
 */
import { HeadersMap, XRPCError } from '@atproto/xrpc'
import { ValidationResult, BlobRef } from '@atproto/lexicon'
import { isObj, hasProp } from '../../../../util'
import { lexicons } from '../../../../lexicons'
import { CID } from 'multiformats/cid'
import * as ComWhtwndBlogDefs from './defs'

export interface QueryParams {
  author: string
}

export type InputSchema = undefined

export interface OutputSchema {
  post: ComWhtwndBlogDefs.BlogEntry[]
  [k: string]: unknown
}

export interface CallOptions {
  signal?: AbortSignal
  headers?: HeadersMap
}

export interface Response {
  success: boolean
  headers: HeadersMap
  data: OutputSchema
}

export function toKnownErr(e: any) {
  return e
}